lmforge/requirements.txt
absl-py==2.1.0
accelerate==1.10.1
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
bert-score==0.3.13
bitsandbytes
branca==0.8.0
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cryptography==43.0.3
ctransformers==0.2.27
cycler==0.12.1
datasets==3.1.0
dill==0.3.8
distlib==0.3.9
distro==1.9.0
Django==4.2.23
django-cors-headers==4.6.0
djangorestframework==3.15.2
docker-pycreds==0.4.0
et_xmlfile==2.0.0
eval_type_backport==0.2.2
evaluate==0.4.3
exceptiongroup==1.2.2
filelock==3.16.1
folium==0.18.0
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.26.2
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
lightning-utilities==0.11.9
lxml==5.3.0
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.2.1
nltk==3.9.1
numpy==2.0.2
openai==1.61.1
opencage==3.0.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pdfminer.six==20231228
pdfplumber==0.11.4
peft==0.17.1
phonenumbers==8.13.50
pillow==11.0.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pyarrow==18.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyMySQL==1.1.1
pyparsing==3.2.1
PyPDF2==3.0.1
pypdfium2==4.30.0
# RAG backend dependencies
pgvector>=0.2.0
langchain-text-splitters>=0.3.0
chromadb>=0.6.9
scikit-learn>=1.3.0
pyproject-toml==0.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rouge_score==0.1.2
rpds-py==0.21.0
safetensors==0.4.5
scikit-learn==1.6.0
scipy
sentence-transformers==4.0.2
sentencepiece==0.2.0
sentry-sdk==2.19.0
setproctitle==1.3.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.2
sympy==1.13.1
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.3
toml==0.10.2
tomli==2.2.1
torch
torchaudio
torchvision
tqdm==4.67.0
transformers==4.46.3
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.27.1
wandb==0.18.7
xxhash==3.5.0
xyzservices==2024.9.0
yarl==1.17.2
zipp==3.21.0
